cloud-asset-inventory/requirements.txt
functions-framework==3.*
google-cloud-bigquery
google-auth